docker safeguards

+++ cicd/scripts/stop-container.sh
@@ -1,9 +1,11 @@
 #!/bin/sh
 
-if [[ $(docker inspect dillo_bot | grep "Running" | grep true) ]]; then
-    docker kill dillo_bot
+if [[ $(docker ps -a | grep dillo_bot) ]]; then
+    if [[ $(docker inspect dillo_bot | grep "Running" | grep true) ]]; then
+        docker kill dillo_bot
 
-    docker rm dillo_bot
+        docker rm dillo_bot
+    fi
 fi
 
 exit
